requirements.txt
# Langchain Core
langchain~=0.3.26
langchain-core~=0.3.66

# OpenAI
langchain-openai~=0.3.24
openai

# Anthropic
langchain-anthropic~=0.3.15

# Google
langchain-google-genai~=2.0.10
google-generativeai

# HF
langchain-huggingface
transformers
huggingface-hub

# Environment
python-dotenv~=1.1.0

# Machine Learning
numpy~=2.3.0
scikit-learn~=1.7.0

# Require to run `TinyLlama/TinyLlama-1.1B-Chat-v1.0` HF locally
torchvision

# Require to run `sentence-transformers/all-MiniLM-L6-v2` HF locally
sentence-transformers

# For building UI
streamlit~=1.46.0

# For types and structured output
pydantic~=2.11.7

# For text splitting
langchain-text-splitters~=0.3.8
langchain-community~=0.3.26
langchain-experimental~=0.3.4